Will leave google drive on hold and focus on mediafire.
